backend/requirements.txt
Flask
flask-cors
requests
certifi
Pillow
pycryptodome
gunicorn